revert(side-nav): adds side nav cards
This reverts commit 968f70217a5a76f0c802db4f6402b02d35966388.
