+++ projects/canopy/src/lib/card/card.component.scss
@@ -1,5 +1,7 @@
-@import '../../styles/mixins.scss';
-
 .lg-card {
-  @include lg-card;
+  display: block;
+  background-color: var(--card-bg-color);
+  border-radius: var(--border-radius-lg);
+  margin-bottom: var(--grid-gutter);
+  padding: var(--component-padding);
 }
