+++ projects/canopy/src/lib/side-nav/index.ts
@@ -3,7 +3,6 @@ export * from './side-nav-bar-footer/side-nav-bar-footer.component';
 export * from './side-nav-bar-item/side-nav-bar-item.component';
 export * from './side-nav-bar-item-content/side-nav-bar-item-content.component';
 export * from './side-nav-bar-item-heading/side-nav-bar-item-heading.component';
-export * from './side-nav-bar-link/side-nav-bar-link.directive';
 export * from './side-nav-content/side-nav-content.component';
 export * from './side-nav.component';
 export * from './side-nav.module';
