--- projects/canopy/src/lib/side-nav/side-nav-bar-footer/side-nav-bar-footer.component.scss
@@ -1,10 +1,0 @@
-@import '../../../styles/mixins';
-
-.lg-side-nav-bar-footer {
-  display: block;
-  padding: var(--space-xxl) var(--component-padding) var(--component-padding);
-
-  @include lg-breakpoint('lg') {
-    padding: var(--space-xxl) 0 0;
-  }
-}
