+++ projects/canopy/src/lib/side-nav/side-nav-bar-footer/side-nav-bar-footer.component.ts
@@ -8,7 +8,6 @@ import {
 @Component({
   selector: 'lg-side-nav-bar-footer',
   templateUrl: './side-nav-bar-footer.component.html',
-  styleUrls: ['./side-nav-bar-footer.component.scss'],
   encapsulation: ViewEncapsulation.None,
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
