+++ projects/canopy/src/lib/side-nav/side-nav-bar-item/side-nav-bar-item.component.html
@@ -1,7 +1,9 @@
-<div class="lg-side-nav-bar-item__container">
-  <ng-content select="lg-side-nav-bar-item-heading"></ng-content>
-  <ng-content select="lg-side-nav-bar-item-content"></ng-content>
-</div>
-<div class="lg-side-nav-bar-item__icon-container">
-  <lg-icon name="chevron-right" class="lg-font-size-0-8"></lg-icon>
+<div class="lg-side-nav-bar-item__inner">
+  <div class="lg-side-nav-bar-item__container">
+    <ng-content select="lg-side-nav-bar-item-heading"></ng-content>
+    <ng-content select="lg-side-nav-bar-item-content"></ng-content>
+  </div>
+  <div class="lg-side-nav-bar-item__icon-container">
+    <lg-icon name="chevron-right" class="lg-font-size-0-8"></lg-icon>
+  </div>
 </div>
