+++ projects/canopy/src/lib/side-nav/side-nav-bar-item/side-nav-bar-item.component.ts
@@ -1,7 +1,10 @@
 import {
   ChangeDetectionStrategy,
   Component,
+  EventEmitter,
   HostBinding,
+  Input,
+  Output,
   ViewEncapsulation,
 } from '@angular/core';
 
@@ -14,4 +17,20 @@ import {
 })
 export class LgSideNavBarItemComponent {
   @HostBinding('class.lg-side-nav-bar-item') class = true;
+
+  _isActive = false;
+
+  @Input()
+  set isActive(isActive) {
+    if (isActive) {
+      this.activated.emit();
+    }
+
+    this._isActive = isActive;
+  }
+  @HostBinding('class.lg-side-nav-bar-item--selected') get isActiveClass() {
+    return this._isActive;
+  }
+
+  @Output() activated: EventEmitter<void> = new EventEmitter();
 }
