+++ projects/canopy/src/lib/side-nav/side-nav-bar/side-nav-bar.component.html
@@ -1,5 +1,1 @@
-<lg-card lgPadding="none" lgMargin="none">
-  <lg-card-content>
-    <ng-content></ng-content>
-  </lg-card-content>
-</lg-card>
+<ng-content></ng-content>
