+++ projects/canopy/src/styles/mixins.scss
@@ -200,11 +200,3 @@ $breakpoints: (
     }
   }
 }
-
-@mixin lg-card {
-  display: block;
-  background-color: var(--card-bg-color);
-  border-radius: var(--border-radius-lg);
-  margin-bottom: var(--grid-gutter);
-  padding: var(--component-padding);
-}
